Rename the module as well
